dashboard/requirements.txt
Flask
cassandra-driver
hdfs
requests
docker